pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "reviews"
version = "0.1.0"
description = "Sample project for BigQuery review fetching."
authors = [
    { name = "Yara Mubarak", email = "yaramubarak8@gmail.com" }
]
requires-python = ">=3.8"
dependencies = [
    "google-cloud-bigquery==3.17.2",
    "pytest==8.2.1",
    "python-dotenv==1.0.1",
    "deprecated==1.2.14",
    "pytest-asyncio==0.23.5",
    "litellm>=1.26.8",
    "google-adk==1.7.0"
]

[tool.setuptools]
packages = ["reviews"]